Made Hash thread-safe and fixed compareTo with an identity check
DO NOT MERGE

(1) Made Hash thread-safe without needing full synchronization.

All access to `bytes`, `trits` and `hashcode` done via a null check against
an inner class object that is instantiated in a synchronized block with
a double check. `ByteSafe` and `TritSafe` are inner classes and the data
hidden in them is not available until the classes are fully instantiated
(classloading ensures this).

(2) #1 fixes the race condition to `hashcode()` and `equals()` and `bytes()`
    that was proven to be happening and might also have cause the FIFO cache
    overrun.

(3) Fixed `compareTo` to do an identity check and not need to make a new
    `Hash` every time - because Hash extends Indexable and the the indexible
    provided could already be a Hash.

(4) Commented on bounds check. This is easy to fix but so many tests are
    abusing it and they won't pass until the tests are also fixed.

(5) Added a null check in `fullread' because the RocksDB does provide null
    byte[] to read methods sometimes and they can be safely passed over.

(6) Added a 2nd time `read` persistence check. The class should be
    effectively immutable (not possible) but at least check to see if it has
    already been initialized and reject a second initialization.

Signed-off-by: Jim Huang <jserv.tw@gmail.com>